requirements.txt
# flux and sgm repo
aiohttp
anyio
certifi
click
coloredlogs
einops
fastapi
filelock
fire
flatbuffers
fsspec
ftfy
huggingface-hub
idna
imageio
Jinja2
jsonschema
kornia
lightning-utilities
markdown-it-py
mpmath
multidict
networkx
numba
numpy
omegaconf
onnxruntime
open_clip_torch
opencv-python-headless
pillow
platformdirs
protobuf
pydantic
Pygments
PyMatting
pytorch-lightning
PyWavelets
PyYAML
regex
rembg
requests
rich
scikit-image
scipy
starlette
sympy
timm
tokenizers
toml
torch
torchmetrics
torchvision
tqdm
transformers
typer
typing_extensions
urllib3
watchfiles
yarl
git+https://github.com/openai/CLIP.git
python-multipart
open-clip-torch==2.24.0
cuda-python
tensorrt
polygraphy
onnx
onnx_graphsurgeon

# Server
uvicorn
xformers
uvicorn[standard]
peft

# sd-scripts 
accelerate
diffusers
imagesize
voluptuous
bitsandbytes
